Rename App.tsx to component.tsx
