+++ src/components/CreateProject/CreateProjectComponent.tsx
@@ -8,7 +8,7 @@ import {
 import Button from "@material-ui/core/Button";
 import TextField from "@material-ui/core/TextField";
 import { Grid } from "@material-ui/core";
-import { history } from "../App/App";
+import { history } from "../App/component";
 
 export interface CreateProjectProps {
   createProject?: (name: string, languageData: File) => void;
